moved simplelog to shared dir
