+++ cmd/local/conf/autodetect/awse.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/ddcio"
-	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
+	"github.com/dremio/dremio-diagnostic-collector/pkg/simplelog"
 )
 
 func IsAWSEFromJPSOutput(jpsText string) (bool, error) {
