+++ cmd/local/conf/autodetect/pid.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/ddcio"
-	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
+	"github.com/dremio/dremio-diagnostic-collector/pkg/simplelog"
 )
 
 func GetDremioPIDFromText(jpsOutput string) (int, error) {
