+++ cmd/local/ddcio/archive.go
@@ -22,7 +22,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
+	"github.com/dremio/dremio-diagnostic-collector/pkg/simplelog"
 )
 
 func GzipFile(src, dst string) error {
