+++ cmd/local/ddcio/ddcio.go
@@ -25,7 +25,7 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
+	"github.com/dremio/dremio-diagnostic-collector/pkg/simplelog"
 )
 
 func DeleteDirContents(dir string) error {
