+++ cmd/root/collection/sudo.go
@@ -17,7 +17,7 @@ package collection
 
 import (
 	"github.com/dremio/dremio-diagnostic-collector/cmd/root/cli"
-	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
+	"github.com/dremio/dremio-diagnostic-collector/pkg/simplelog"
 )
 
 // Adds the sudo part into the HostExecute call
