+++ pkg/simplelog/logger_test.go
